requirements.txt
python-telegram-bot
gTTS
pydub